routes are now hit on a timer. large structural changes that should have been multiple commits.

+++ response.go
@@ -2,7 +2,7 @@ package main
 
 type Response struct {
     Size       int64
-    Duration   int64
+    Duration   float64
     StatusCode int
     Error      bool
 }
